+++ app/build.gradle
@@ -2,8 +2,9 @@ plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
     id 'kotlin-kapt'
-   id 'com.google.dagger.hilt.android'
-
+    id 'com.google.dagger.hilt.android'
+    id 'androidx.navigation.safeargs'
+    //id 'dagger.hilt.android.plugin'
 
 }
 
@@ -36,9 +37,10 @@ android {
         jvmTarget = '1.8'
     }
 
-    buildFeatures{
+    buildFeatures {
         viewBinding true
     }
+
 
 }
 
@@ -67,12 +69,20 @@ dependencies {
     implementation "androidx.activity:activity-ktx:1.8.0"
 
     //dagger-hilt
-    implementation"com.google.dagger:hilt-android:2.44"
-    kapt 'com.google.dagger:hilt-compiler:2.44'
-    kapt 'androidx.hilt:hilt-compiler:1.0.0'
+    def hilt_version="2.48.1"
+    implementation "com.google.dagger:hilt-android:$hilt_version"
+    kapt "com.google.dagger:hilt-compiler:$hilt_version"
+    kapt "androidx.hilt:hilt-compiler:1.0.0"
+    kapt "org.jetbrains.kotlinx:kotlinx-metadata-jvm:0.5.0"
+
 
     //coil
     implementation("io.coil-kt:coil:1.1.1")
+
+    //navigation
+    def nav_version = "2.7.5"
+    implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
+    implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
 }
 
 kapt {
