+++ app/src/main/java/ru/perelyginva/test/MainActivity.kt
@@ -1,19 +1,13 @@
 package ru.perelyginva.test
 
 import android.os.Bundle
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.LinearLayoutManager
 import dagger.hilt.android.AndroidEntryPoint
-import ru.perelyginva.test.adapter.TvShowAdapter
 import ru.perelyginva.test.databinding.ActivityMainBinding
-import ru.perelyginva.test.viewmodel.TvShowViewModel
 
 @AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
-    private val viewModel: TvShowViewModel by viewModels()
-    private lateinit var tvShowAdapter: TvShowAdapter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -21,51 +15,6 @@ class MainActivity : AppCompatActivity() {
         setContentView(binding.root)
         supportActionBar?.hide()
 
-        setUpRv()
-
     }
 
-
-    private fun setUpRv() {
-        tvShowAdapter = TvShowAdapter()
-
-        binding.recyclerView.apply {
-            adapter = tvShowAdapter
-            layoutManager = LinearLayoutManager(
-                this@MainActivity, LinearLayoutManager.HORIZONTAL,
-                false
-            )
-
-            setHasFixedSize(true)
-        }
-
-        binding.rvEpisodes.apply {
-            adapter = tvShowAdapter
-            layoutManager = LinearLayoutManager(
-                this@MainActivity, LinearLayoutManager.HORIZONTAL,
-                false
-            )
-
-            setHasFixedSize(true)
-        }
-
-
-        binding.rvRecentlyAdded.apply {
-            adapter = tvShowAdapter
-            layoutManager = LinearLayoutManager(
-                this@MainActivity, LinearLayoutManager.HORIZONTAL,
-                false
-            )
-
-            setHasFixedSize(true)
-        }
-
-
-        viewModel.repositoryTvShow.observe(this) { listTvShows ->
-
-            tvShowAdapter.tvShows = listTvShows
-
-        }
-
-    }
 }
