+++ app/src/main/java/ru/perelyginva/test/StartFragment.kt
@@ -1,0 +1,88 @@
+package ru.perelyginva.test
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.activity.viewModels
+import androidx.fragment.app.viewModels
+import androidx.recyclerview.widget.LinearLayoutManager
+import dagger.hilt.android.AndroidEntryPoint
+import ru.perelyginva.test.adapter.TvShowAdapter
+import ru.perelyginva.test.databinding.FragmentStartBinding
+import ru.perelyginva.test.viewmodel.TvShowViewModel
+
+@AndroidEntryPoint
+class StartFragment : Fragment() {
+
+    private var _binding: FragmentStartBinding ? = null
+    private val binding: FragmentStartBinding
+        get() = _binding ?: throw RuntimeException("FragmentStartBinding == null")
+    private val viewModel: TvShowViewModel by viewModels()
+    private lateinit var tvShowAdapter: TvShowAdapter
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+
+        _binding = FragmentStartBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setUpRv()
+    }
+
+
+    private fun setUpRv() {
+        tvShowAdapter = TvShowAdapter()
+
+        binding.recyclerView.apply {
+            adapter = tvShowAdapter
+            layoutManager = LinearLayoutManager(
+                context, LinearLayoutManager.HORIZONTAL,
+                false
+            )
+
+            setHasFixedSize(true)
+        }
+
+        binding.rvEpisodes.apply {
+            adapter = tvShowAdapter
+            layoutManager = LinearLayoutManager(
+                context, LinearLayoutManager.HORIZONTAL,
+                false
+            )
+
+            setHasFixedSize(true)
+        }
+
+
+        binding.rvRecentlyAdded.apply {
+            adapter = tvShowAdapter
+            layoutManager = LinearLayoutManager(
+                context, LinearLayoutManager.HORIZONTAL,
+                false
+            )
+
+            setHasFixedSize(true)
+        }
+
+
+        viewModel.repositoryTvShow.observe(viewLifecycleOwner) { listTvShows ->
+
+            tvShowAdapter.tvShows = listTvShows
+
+        }
+
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding == null
+    }
+}
