+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,6 @@
     <string name="up_coming">Up Coming</string>
     <string name="tv_shows">Tv Shows</string>
     <string name="textview">TextView</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
