+++ build.gradle
@@ -2,7 +2,16 @@
 
 buildscript {
     ext {
-        kotlin_version = '1.6.10'
+        kotlin_version = '1.7.0'
+    }
+    repositories {
+        // other repositories...
+        mavenCentral()
+    }
+
+    dependencies {
+        def nav_version = "2.7.5"
+        classpath("androidx.navigation:navigation-safe-args-gradle-plugin:2.7.5")
 
     }
 }
@@ -10,6 +19,7 @@ buildscript {
 plugins {
     id 'com.android.application' version '8.1.2' apply false
     id 'org.jetbrains.kotlin.android' version '1.7.0' apply false
-    id 'com.google.dagger.hilt.android' version '2.44' apply false
+    id 'com.google.dagger.hilt.android' version '2.48.1' apply false
+    id 'androidx.navigation.safeargs' version '2.7.5' apply false
 
 }
